Add SuperCookingGuildDoor bot
Someone asked me to make them a bot that just goes in and out of the cooking guild.
I have no idea why, but I'm getting paid to do it.

I had to strip much of the regular bot code that I use since this bot is so simple. I may
be able to avoid duplicating so much code (since I had to make new, stripped down
versions of the files) at a later date.

git-svn-id: https://svn.runemate.com/svn/superbotter@31 a474b641-a148-44e9-9bf8-48c4f81c8702