Make walking twice as fast
This update made walking twice as fast (or half as slow) by minimising unnecessary Path generation. Why generate a RegionPath if the destination is not in the Region? The two new travel methods (the old ones were deleted) now check for that.

The travelTo methods have been replaced by the getPathTo methods. This allows the Path to be saved and reused while the bot still needs to walk along it instead of being regenerated every time the bot has to move. This means that after the path is made once, it never needs to be remade until the bot has to go to a different area. This increases travel speeds greatly.

git-svn-id: https://svn.runemate.com/svn/superbotter@36 a474b641-a148-44e9-9bf8-48c4f81c8702