Fix bug when walking to bank
I fixed the bug that occured when the bot's inventory was full and it had to go to the bank. It ended up just going to the botArea forever.

I added the Al Kharid mining site to SuperMiner.

Sorted ore/fish alphabetically.

Made NonMenuAction use less redundant variables. It still needs work, but it's better now.

git-svn-id: https://svn.runemate.com/svn/superbotter@37 a474b641-a148-44e9-9bf8-48c4f81c8702