Remade dropping and banking systems
The dropping and banking systems now make use of a new class called
RequiredItems. This class contains methods for adding items to it and
getting the items that are not present in the Inventory, and many more.
If the player runs out of the RequiredItems it either goes back to the
bank to get more or it stops the bot if banking is disabled.

I also made it so that each Task selects its own loopDelay and I fixed
some error handling so that when the bot crashes its more user friendly.

git-svn-id: https://svn.runemate.com/svn/superbotter@39 a474b641-a148-44e9-9bf8-48c4f81c8702