Add the Dwarven Mine to SuperMiner and fix bugs
Added support for the Dwarven Mine to SuperMiner. It currently does NOT
handle the aggressive scorpions in the mine yet. It uses the Falador
East bank and support has been added for it.

Added support for the Shattered Worlds bank. It is not used until I can
find a way to check if the bot is a member or not.

Fixed bug in api/ui/Config that caused the bot to crash when it tried to
stop the bot because Environment.getBot() throws an error when used
outside of the bot thread.

git-svn-id: https://svn.runemate.com/svn/superbotter@40 a474b641-a148-44e9-9bf8-48c4f81c8702