Replace RequiredItems with ProtectedItems
RequiredItems has been expanded and renamed to ProtectedItems, which now
tracks the "status" of each item put inside. The status decides if the
item is required for the bot, wanted for the bot, or just shouldn't be
dropped (comments in the code explain it better).

All bots now support strange rocks. They are added to the RequiredItems
of each bot with amount 0 and status 0.

git-svn-id: https://svn.runemate.com/svn/superbotter@41 a474b641-a148-44e9-9bf8-48c4f81c8702