Fix bugs caused by last update
Last update I used SVN to only commit parts of some files that I edited,
and to not commit some files altogether.
SuperFisher/ui/ConfigController.java was excluded from the commit and I
forgot that some of the edits to that file ensured that RequiredItems
became ProtectedItems. It should now be fixed.

git-svn-id: https://svn.runemate.com/svn/superbotter@42 a474b641-a148-44e9-9bf8-48c4f81c8702