Add urn support and fix bugs
Added Urn support to the bot using a new Urn Task. The Config screen has
been modified to add support for urns as well.

The statuses in ProtectedItems is now an enum, Status. A new status,
HELD was also added for urn support.

The ConfigControllers have been simplified. It needs to be rewritten
eventually but it has been cleaned up a bit.

I fixed a bug that was called when the first ActionBar action for an
item was not drop. An example is gems, the first action is Craft so the
bot would get stuck trying to craft gems instead of dropping them.

Fixed a few bugs on start related to using Environment.getBot() outside
of the main bot thread.

git-svn-id: https://svn.runemate.com/svn/superbotter@43 a474b641-a148-44e9-9bf8-48c4f81c8702