Add multimining/multifishing support
This is a pretty huge update that I worked on a bit last August, and I
finished it the past few days. It should be split into multiple commits,
but I just want to get a commit finished.

InteractableItems - new system for managing items that the bot needs to
collect. This allows for multimining support and it hopefully makes
things a bit cleaner.

InfoUpdate - not super sure why I started working on this last year, but
I finished it now for a release. It may be removed in the future.

Upcoming updates: More OOP (SuperBot superclass and new ConfigController
system to handle the lots of repeated code between the two bots),
conversion to TreeBot possibly.

git-svn-id: https://svn.runemate.com/svn/superbotter@44 a474b641-a148-44e9-9bf8-48c4f81c8702