Make new API classes to decrease repeated code
New file: SuperBot
SuperBot handles all of the repeated code between SuperMiner and
SuperFisher. Now, those classes just provide all of the specific stuff,
namely the locations and the items to collect. This stuff also required
more stuff to be generified between the two bots, like the
ConfigControllers. The new ConfigController class just handles the
repeated code between the old Config controllers. This required the new
Location class and some expansions to InteractableItems (which is now
CollectableItems), namely making it hold its own ProtectedItems to fully
handle collecting multiple items at once with their own requirements.

What all this means for the end user is that they can collect multiple
types of items at once, like sardines (bait fish) and shrimp (net fish),
and when one of them can't be collected anymore because it's out of a
required item, it continues to collect the other ones.

Hopefully this is one of the last really big updates for a while.
We'll see.

git-svn-id: https://svn.runemate.com/svn/superbotter@46 a474b641-a148-44e9-9bf8-48c4f81c8702