Fix dropping
Dropping was reworked. It should work now and it might be a bit faster.
There is a slight bug when it drags an item to the Action Bar, it drags
it twice for some reason. I'll put it on the GitHub Issues page but I
can't really figure out why it happens.

Dropping also supports the new ProtectedItems in CollectableItems that
was introduced a couple commits ago.

git-svn-id: https://svn.runemate.com/svn/superbotter@48 a474b641-a148-44e9-9bf8-48c4f81c8702