Add lots of mines and fix teleport bug
The location of all of the F2P mines have been added to SuperMiner. Some
of them can't be accessed since they're not currently in the RuneMate
web, so I have commented them out and I'll work something else out for
them. The Karamja mine and the Crandor mines can't be accessed by F2P
players, but they can be by P2P players so I left them in for now. The
Mining Guild can only be accessed through the ladders in Falador, not
through the Dwarven Mine door. All of these will eventually be fixed and
full support will be added for a v1.0.0 release.

A bug happened in which the bot would keep on trying to teleport while
it was already trying to teleport. This would result in an infinite
loop. This has been fixed in both bots.

git-svn-id: https://svn.runemate.com/svn/superbotter@50 a474b641-a148-44e9-9bf8-48c4f81c8702