Fix bug caused when removing ProtectedItems
I forgot a rule about ArrayLists and it made
ArrayIndexOutOfBoundsExceptions get thrown if the user selected an urn
at first, then selected "no urn". It wouldn't remove all of the
ProtectedItems it generated and this caused an error in the Urn class
when it tried to get the name of the urn. It's fixed now.

git-svn-id: https://svn.runemate.com/svn/superbotter@56 a474b641-a148-44e9-9bf8-48c4f81c8702